fix: Fixes issues described in #105
Moves prevent default to pointermove as opposed to touchstart to avoid issues with touchstart clicking into a draggable item. Makes improvements from switching to desktop to mobile (so it'll be reactive and page reload isn't neccessary). Fixes issue with longPress where it wasn't actually being validated that it had been pressed long enough